This pull request merges the changes from the dev-2 branch into the main
branch.

**Branch/Task Summary:**
- api_6: Add unit tests for repository files
- api_7: Create entity unit tests 
- api_8: Add unit tests for service layer 
- api-9: Create controller layer
- api-10: Integrate Spring Security

Each task represents a significant development milestone, covering
different aspects of the project's architecture and functionality.

**Changes Made:**
- Unit tests were added for repository files, entity layer, service
layer, and controller layer.
- Entity classes, repository interfaces, service classes, and controller
classes were implemented or modified as necessary.
- Spring Security was integrated into the project, providing user
authentication and authorization features.
- Database initialization and security configuration were added.

This pull request consolidates all the changes made in the dev-2 branch
and prepares them for deployment to the main branch, ensuring that the
latest features and improvements are available in the production
environment.

No conflicts were encountered during the merge process.

+++ src/main/java/online/lcelectronics/api/entities/Client.java
@@ -18,7 +18,7 @@ public class Client {
     // Primary key for the Client table
     @Id
     @NotNull(message = "Identity card cannot be null")
-    private Integer identityCard;
+    private Long identityCard;
 
     // Name of the client
     @NotBlank(message = "Name cannot be blank")
@@ -26,7 +26,7 @@ public class Client {
 
     // Phone number of the client
     @NotNull(message = "Phone number cannot be null")
-    private Integer phone;
+    private Long phone;
 
     // Address of the client
     @NotBlank(message = "Address cannot be blank")
